style: move all global enums to root package
